+++ game_patch/multi/network.cpp
@@ -1424,7 +1424,7 @@ CodeInjection process_obj_update_health_armor_injection{
         if (!spectated || spectated == rf::local_player || multi_spectate_is_freelook())
             return;
         if (entity == rf::entity_from_handle(spectated->entity_handle))
-            rf::local_screen_flash(rf::local_player, 255, 0, 0, 128);
+            player_damage_feedback();
     },
 };